dom/tests/mochitest/webapps/test_list_api.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/
   - https://bugzilla.mozilla.org/show_bug.cgi?id=741549
-->

<window title="Mozilla Bug 741549"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript" src="apphelper.js"/>
  <script type="application/javascript" src="jshelper.js"/>
  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=741549"
     target="_blank">Mozilla Bug 741549</a>
  </body>

<script type="text/javascript">
<![CDATA[

var expectedMethods = ['navigator.mozApps.QueryInterface',
                       'navigator.mozApps.getInstalled',
                       'navigator.mozApps.getSelf',
                       'navigator.mozApps.install',
                       'navigator.mozApps.mgmt.QueryInterface',
                       'navigator.mozApps.mgmt.addEventListener',
                       'navigator.mozApps.mgmt.dispatchEvent',
                       'navigator.mozApps.mgmt.getAll',
                       'navigator.mozApps.mgmt.removeEventListener'];
var actualMethods = [];

actualMethods = actualMethods.concat(iterateMethods("navigator.mozApps.", navigator.mozApps, ['setApplicationChooser', 'setRepoOrigin', 'setMockResponse']));
actualMethods = actualMethods.concat(iterateMethods("navigator.mozApps.mgmt.", navigator.mozApps.mgmt));
actualMethods = actualMethods.sort();
ok(JSON.stringify(actualMethods) == JSON.stringify(expectedMethods), actualMethods);
tearDown();
]]>

</script> 

</window>
